# public/sdk/openapi.yaml
openapi: "3.0.3"
info:
  title: Print Hub API
  description: |
    Centralized print management middleware API.

    Print Hub decouples client applications from physical printers. Client apps
    submit print jobs via this REST API, and Print Hub routes them to the
    correct print agent (TrayPrint) for execution.

    ## Authentication

    Two authentication modes are supported:

    | Role          | Method    | Header                             |
    |---------------|-----------|------------------------------------|
    | Client App    | API Key   | `X-API-Key: {your_api_key}`        |
    | Print Agent   | Bearer    | `Authorization: Bearer {agent_key}`|

    Rate limits:
    - Client API: 60 requests/minute
    - Agent API:  120 requests/minute
  version: "1.0.0"
  contact:
    name: Print Hub Support
    url: https://print-hub.example.com/support
  license:
    name: Proprietary
servers:
  - url: https://print-hub.example.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /test:
    get:
      summary: Test API connection
      description: Verify that the API key is valid and the server is reachable.
      operationId: testConnection
      tags: [Connection & Health]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Connection successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestConnectionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /branches:
    get:
      summary: List all branches
      description: Returns a list of all active company branches.
      operationId: listBranches
      tags: [Discovery]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Branch list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchListResponse"
  /agents/online:
    get:
      summary: List online print agents
      description: Returns agents currently connected and accepting jobs.
      operationId: listOnlineAgents
      tags: [Discovery]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: branch_code
          in: query
          required: false
          schema:
            type: string
          description: Filter by branch code
      responses:
        "200":
          description: Online agent list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnlineAgentListResponse"
  /queues:
    get:
      summary: List print queues
      description: Returns queues configured per branch.
      operationId: listQueues
      tags: [Discovery]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: branch_code
          in: query
          required: false
          schema:
            type: string
          description: Filter by branch code
        - name: detailed
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Include full printer configuration details
      responses:
        "200":
          description: Queue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueListResponse"

  # =====================================================================
  # Templates
  # =====================================================================
  /templates:
    get:
      summary: List all print templates
      description: Returns all available template definitions.
      operationId: listTemplates
      tags: [Templates]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResponse"
  /templates/{name}:
    get:
      summary: Get template details
      description: Returns detailed information for a specific template including field positions.
      operationId: getTemplate
      tags: [Templates]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Template name (URL-encoded)
      responses:
        "200":
          description: Template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /templates/{name}/schema:
    get:
      summary: Get template data schema
      description: Returns the required data schema for a template.
      operationId: getTemplateSchema
      tags: [Templates]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Template name
      responses:
        "200":
          description: Template schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateSchemaResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # =====================================================================
  # Data Schemas
  # =====================================================================
  /schema:
    post:
      summary: Register or update a data schema
      description: Register a new schema or update an existing one for template data binding.
      operationId: registerSchema
      tags: [Data Schemas]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSchemaRequest"
      responses:
        "200":
          description: Schema registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterSchemaResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /schemas:
    get:
      summary: List all registered schemas
      operationId: listSchemas
      tags: [Data Schemas]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Schema list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaListResponse"
  /schema/{name}/versions:
    get:
      summary: Get schema version history
      operationId: schemaVersions
      tags: [Data Schemas]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Schema name
      responses:
        "200":
          description: Schema versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaVersionsResponse"

  # =====================================================================
  # Printing
  # =====================================================================
  /print:
    post:
      summary: Submit a print job
      description: >
        Print a document using either a template (with data fields) or a raw
        base64-encoded PDF. When using a template, the server renders the
        document using the template's layout and the provided data. When
        providing `base64_pdf`, the PDF is sent directly to the printer.
      operationId: submitPrintJob
      tags: [Printing]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrintJobRequest"
      responses:
        "200":
          description: Job queued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJobResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /print/batch:
    post:
      summary: Batch print
      description: Submit up to 50 print jobs in a single request.
      operationId: batchPrint
      tags: [Printing]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchPrintRequest"
      responses:
        "200":
          description: Jobs queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchPrintResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /preview:
    post:
      summary: Preview a template as PDF
      description: Generate a PDF preview of a template with the provided data without sending it to a printer.
      operationId: previewPrint
      tags: [Printing]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewRequest"
      responses:
        "200":
          description: PDF binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "422":
          $ref: "#/components/responses/ValidationError"

  # =====================================================================
  # Job Management
  # =====================================================================
  /jobs/{job_id}:
    get:
      summary: Get job status
      operationId: jobStatus
      tags: [Job Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Print job ID
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Cancel a pending job
      operationId: cancelJob
      tags: [Job Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Print job ID
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCancelResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # =====================================================================
  # Document Management
  # =====================================================================
  /documents/upload:
    post:
      summary: Upload a document
      operationId: uploadDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (PDF, TXT, images)
                label:
                  type: string
                  description: Optional label for the document
              required:
                - file
      responses:
        "200":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /documents:
    get:
      summary: List documents
      operationId: listDocuments
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
  /documents/{id}:
    get:
      summary: Get document metadata
      operationId: getDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Document ID
      responses:
        "200":
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a document
      operationId: deleteDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Document deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{id}/preview:
    get:
      summary: Preview a document
      operationId: previewDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document preview
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{id}/download:
    get:
      summary: Download a document
      operationId: downloadDocument
      tags: [Document Management]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"

  # =====================================================================
  # Connector Registry
  # =====================================================================
  /connectors:
    get:
      summary: List all connectors
      description: Returns all connectors registered for the authenticated client app.
      operationId: listConnectors
      tags: [Connector Registry]
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Connector list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorListResponse"
    post:
      summary: Register a new connector
      description: Create a new data-source connector definition.
      operationId: registerConnector
      tags: [Connector Registry]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterConnectorRequest"
      responses:
        "201":
          description: Connector created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorDetailResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /connectors/{id}:
    put:
      summary: Update a connector
      description: Update an existing connector's configuration.
      operationId: updateConnector
      tags: [Connector Registry]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Connector UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnectorRequest"
      responses:
        "200":
          description: Connector updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      summary: Delete a connector
      description: Remove a connector definition.
      operationId: deleteConnector
      tags: [Connector Registry]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Connector UUID
      responses:
        "200":
          description: Connector deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Connector deleted successfully."
        "404":
          $ref: "#/components/responses/NotFound"
  /connectors/{id}/test:
    post:
      summary: Test a connector connection
      description: Sends a HEAD request to the connector's configured URL to verify connectivity.
      operationId: testConnector
      tags: [Connector Registry]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Connector UUID
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorTestResponse"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Client app API key obtained from Print Hub admin panel
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "NOT_FOUND"
            message:
              type: string
              example: "Resource not found"
...
    TestConnectionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Connected successfully."
        app_name:
          type: string
          example: "Print Hub"
        version:
          type: string
          example: "1.0.0"
        agents:
          type: integer
          example: 3
...
    BranchListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Branch"
...
    OnlineAgentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/OnlineAgent"
...
    QueueListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Queue"
...
    TemplateListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Template"
...
    TemplateDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/TemplateDetail"
...
    TemplateSchemaResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/TemplateSchema"
...
    RegisterSchemaRequest:
      type: object
      required:
        - schema_name
        - fields
      properties:
        schema_name:
          type: string
          example: "invoice_sewa"
        label:
          type: string
          example: "Sewa Invoice"
        fields:
          type: object
          description: Field definitions (label, type, required)
        tables:
          type: object
          description: Table definitions with columns
    RegisterSchemaResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    SchemaListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/SchemaSummary"
...
    SchemaVersionsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              version:
                type: integer
              created_at:
                type: string
                format: date-time
              schema_data:
                type: object

    # --- Printing ---
    PrintJobRequest:
      type: object
      properties:
        template:
          type: string
          description: Template name (required for template-based printing)
          example: "invoice_sewa"
        data:
          type: object
          description: Field values matching the template schema
        base64_pdf:
          type: string
          description: Base64-encoded PDF content (for raw PDF printing)
        reference_id:
          type: string
          description: Your internal reference ID
          example: "INV-001"
        branch_code:
          type: string
          description: Target branch code
          example: "SDP-SBY"
        printer_name:
          type: string
          description: Target printer name (required when no template)
        options:
          type: object
          description: Print options (copies, duplex, color_mode, etc.)
          properties:
            copies:
              type: integer
              default: 1
            duplex:
              type: string
              enum: [none, two-sided-long-edge, two-sided-short-edge]
            color_mode:
              type: string
              enum: [color, monochrome]
            orientation:
              type: string
              enum: [portrait, landscape]
            paper_size:
              type: string
            tray_source:
              type: string
            collate:
              type: boolean
            fit_to_page:
              type: boolean
            eco_mode:
              type: boolean
              description: Enable eco-friendly printing (grayscale, N-up, duplex)
            grayscale_force:
              type: boolean
            pages_per_sheet:
              type: integer
              enum: [1, 2, 4, 6, 8, 9, 16]
            remove_images:
              type: boolean
    PrintJobResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        status:
          type: string
          example: "queued"
        job_id:
          type: string
          example: "abc-123-def"
        agent:
          type: string
          example: "PC-SBY-01"
        printer:
          type: string
          example: "HP LaserJet M404"
    BatchPrintRequest:
      type: object
      required:
        - jobs
      properties:
        jobs:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/PrintJobRequest"
    BatchPrintResponse:
      type: object
      properties:
        success:
          type: boolean
        results:
          type: array
          items:
            $ref: "#/components/schemas/PrintJobResponse"
        errors:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              error:
                type: string
    PreviewRequest:
      type: object
      required:
        - template
        - data
      properties:
        template:
          type: string
        data:
          type: object
        options:
          type: object

    # --- Job Management ---
    JobStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/JobDetail"
...
    JobCancelResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Job cancelled"

    # --- Document Management ---
    DocumentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Document"
...
    DocumentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        meta:
          type: object
          properties:
            current_page:
              type: integer
            last_page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer

    # --- Connector Registry ---
...
    ConnectorListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            connectors:
              type: array
              items:
                $ref: "#/components/schemas/Connector"
    ConnectorDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            connector:
              $ref: "#/components/schemas/Connector"
    RegisterConnectorRequest:
      type: object
      required:
        - name
        - type
        - config
      properties:
        name:
          type: string
          maxLength: 255
          example: "SDP Finance ERP"
        type:
          type: string
          enum: [api, webhook, odoo, custom]
          example: "api"
        config:
          type: object
          description: Connection configuration
          example:
            endpoint_url: "https://erp.example.com/api/v1"
            auth_type: "basic"
        icon:
          type: string
          nullable: true
          example: "🔌"
    UpdateConnectorRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          example: "SDP Finance ERP (Updated)"
        type:
          type: string
          enum: [api, webhook, odoo, custom]
          example: "api"
        config:
          type: object
          description: Connection configuration
        icon:
          type: string
          nullable: true
        is_active:
          type: boolean
    ConnectorTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            connector_id:
              type: string
              format: uuid
            success:
              type: boolean
              description: Whether the test connection succeeded
            message:
              type: string
              example: "Connected successfully (HTTP 200)."
            latency_ms:
              type: integer
              nullable: true
              description: Response time in milliseconds
            last_test_at:
              type: string
              format: date-time
              nullable: true